+++ libs/prisma/tsconfig.lib.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "declaration": true,
+    "outDir": "../dist/libs/prisma"
+  },
+  "include": ["src/**/"],
+  "exclude": ["node_modules", "dist", "test", "**/*spec.ts"]
+}
